Share one string interner and one layer-function table
The byte-identical matklad interners in ipc2581 and gerberx2 move to
a new pcb-intern crate that both re-export. The six-variant copper
layer-function set, previously repeated in four match tables across
pcb-ipc2581-tools, becomes layers::is_copper, and the layer-function
to render-role mapping moves from geometry::render into the shared
layers module.
