+++ Cargo.toml
@@ -25,6 +25,7 @@ mimalloc = "0.1"
 similar = "3"
 
 ipc2581 = { path = "crates/ipc2581", version = "0.3.3" }
+pcb-intern = { path = "crates/pcb-intern", version = "0.1.0" }
 gerberx2 = { path = "crates/gerberx2", version = "0.1.2" }
 pcb-ir = { path = "crates/pcb-ir", version = "0.3.82" }
 pcb-command-runner = { path = "crates/pcb-command-runner" }
