+++ crates/gerberx2/src/lib.rs
@@ -1,11 +1,10 @@
 pub mod from_artwork;
 pub mod geometry;
-mod intern;
 mod parse;
 pub mod types;
 pub mod write;
 
-pub use intern::{Interner, Symbol};
+pub use pcb_intern::{Interner, Symbol};
 pub use types::*;
 pub use write::{
     AttributeValue, GerberLayer, WriterAperture, WriterApertureMacro, WriterApertureTemplate,
