+++ crates/ipc2581/Cargo.toml
@@ -12,6 +12,7 @@ categories = ["parsing", "hardware-support"]
 
 [dependencies]
 hex = { workspace = true }
+pcb-intern = { workspace = true }
 quick-xml = { workspace = true }
 md-5 = { workspace = true }
 base64 = { workspace = true }
