+++ crates/pcb-intern/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "pcb-intern"
+version = "0.1.0"
+edition.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+description = "Small string interner shared by the pcb format parsers"
+license = "MIT OR Apache-2.0"
+keywords = ["interner", "string", "pcb"]
+categories = ["data-structures"]
+
+[dependencies]
